+++ lib/ratting_small_tag.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class RattingSmallTag extends StatelessWidget {
+  const RattingSmallTag({super.key, required this.data, this.onPressed});
+
+  final String data;
+  final VoidCallback? onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onPressed,
+      child: Container(
+        padding: const EdgeInsets.only(
+          left: 5,
+          right: 5,
+          bottom: 3,
+          top: 3,
+        ),
+        decoration: BoxDecoration(
+          color: Theme.of(context).primaryColorLight,
+          borderRadius: const BorderRadius.only(
+            topLeft: Radius.circular(3),
+            bottomLeft: Radius.circular(3),
+            bottomRight: Radius.circular(3),
+          ),
+        ),
+        child: Text(
+          data,
+          style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                color: Colors.white,
+                fontWeight: FontWeight.w600,
+              ),
+        ),
+      ),
+    );
+  }
+}