+++ lib/status_tag.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class StatusTag extends StatelessWidget {
+  const StatusTag({super.key, required this.data, this.onPressed});
+
+  final String data;
+  final VoidCallback? onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onPressed,
+      child: Container(
+        padding: const EdgeInsets.only(
+          left: 5,
+          right: 5,
+          top: 3,
+          bottom: 3,
+        ),
+        decoration: BoxDecoration(
+          color: Theme.of(context).primaryColor,
+          borderRadius: const BorderRadius.all(Radius.circular(10)),
+        ),
+        child: Text(
+          data,
+          style: Theme.of(context)
+              .textTheme
+              .bodySmall
+              ?.copyWith(color: Colors.white),
+        ),
+      ),
+    );
+  }
+}