+++ lib/tag_builder.dart
@@ -1,5 +1,9 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'app_tag.dart';
+export 'chip_tag.dart';
+export 'gray_tag.dart';
+export 'offer_tag.dart';
+export 'out_line_tag.dart';
+export 'primary_icon_tag.dart';
+export 'primary_tag.dart';
+export 'ratting_small_tag.dart';
+export 'status_tag.dart';